[CIAS30-3822] Group autofinish controls in a single component
