+++ app/components/FormikSwitchInput/index.tsx
@@ -1,5 +1,7 @@
 import React, { memo } from 'react';
-import { useField } from 'formik';
+import { useField, useFormikContext } from 'formik';
+
+import useDidUpdateEffect from 'utils/useDidUpdateEffect';
 
 import FormikControlLayout from 'components/FormikControlLayout';
 import { Switch, Props as SwitchProps } from 'components/Switch';
@@ -7,17 +9,32 @@ import { Switch, Props as SwitchProps } from 'components/Switch';
 export type Props = {
   formikKey: string;
   columnStyleProps?: Record<string, unknown>;
+  submitOnChange?: boolean;
+  onChange?: (value: boolean) => void;
 } & Omit<SwitchProps, 'onToggle' | 'onChange' | 'nativeChangeHandler' | 'id'>;
 
 const FormikSwitchInput = ({
   formikKey,
   children,
   columnStyleProps,
+  submitOnChange,
+  onChange,
   ...props
 }: Props) => {
+  const { submitForm } = useFormikContext();
+
   const [field, meta] = useField(formikKey);
   const { error, touched } = meta;
   const { value, ...fieldProps } = field;
+
+  useDidUpdateEffect(() => {
+    if (onChange) {
+      onChange(value);
+    }
+    if (submitOnChange) {
+      submitForm();
+    }
+  }, [value]);
 
   return (
     <FormikControlLayout
